% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marianes_htseq_DESeq.R
\docType{data}
\name{marianes_htseq_DESeq}
\alias{marianes_htseq_DESeq}
\title{DESeq results from regions of the Drosophila midgut}
\format{
An object of class \code{DESeqDataSet} with 17559 rows and 30 columns.
}
\usage{
marianes_htseq_DESeq
}
\description{
DESeqDataSet containing analysis of RNA-seq readcounts from regions of the Drosophila midgut
}
\seealso{
\link{marianes_htseq_metadata} for counts data and information about samples

\code{\link[=DESeq]{DESeq()}}
}
\keyword{datasets}
