% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marianes_htseq.R
\docType{data}
\name{marianes_htseq_metadata}
\alias{marianes_htseq_metadata}
\alias{marianes_htseq_am1}
\alias{marianes_htseq_am2}
\alias{marianes_htseq_am3}
\alias{marianes_htseq_am4}
\alias{marianes_htseq_am5}
\alias{marianes_htseq_am6}
\alias{marianes_htseq_am7}
\alias{marianes_htseq_am8}
\alias{marianes_htseq_am9}
\alias{marianes_htseq_am10}
\alias{marianes_htseq_am11}
\alias{marianes_htseq_am12}
\alias{marianes_htseq_am13}
\alias{marianes_htseq_am14}
\alias{marianes_htseq_am15}
\alias{marianes_htseq_am16}
\alias{marianes_htseq_am17}
\alias{marianes_htseq_am18}
\alias{marianes_htseq_am19}
\alias{marianes_htseq_am20}
\alias{marianes_htseq_am21}
\alias{marianes_htseq_am22}
\alias{marianes_htseq_am23}
\alias{marianes_htseq_am24}
\alias{marianes_htseq_am25}
\alias{marianes_htseq_am26}
\alias{marianes_htseq_am27}
\alias{marianes_htseq_am28}
\alias{marianes_htseq_am29}
\alias{marianes_htseq_am30}
\title{HTSeq readcounts from regions of the Drosophila midgut}
\format{
An object of class \code{data.frame} with 30 rows and 3 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.

An object of class \code{data.frame} with 17564 rows and 2 columns.
}
\usage{
marianes_htseq_metadata

marianes_htseq_am1

marianes_htseq_am2

marianes_htseq_am3

marianes_htseq_am4

marianes_htseq_am5

marianes_htseq_am6

marianes_htseq_am7

marianes_htseq_am8

marianes_htseq_am9

marianes_htseq_am10

marianes_htseq_am11

marianes_htseq_am12

marianes_htseq_am13

marianes_htseq_am14

marianes_htseq_am15

marianes_htseq_am16

marianes_htseq_am17

marianes_htseq_am18

marianes_htseq_am19

marianes_htseq_am20

marianes_htseq_am21

marianes_htseq_am22

marianes_htseq_am23

marianes_htseq_am24

marianes_htseq_am25

marianes_htseq_am26

marianes_htseq_am27

marianes_htseq_am28

marianes_htseq_am29

marianes_htseq_am30
}
\description{
Fill in more documentation
}
\section{Functions}{
\itemize{
\item \code{marianes_htseq_am1}: Individual replicate from the  a1-3 region

\item \code{marianes_htseq_am2}: Individual replicate from the  a1-3 region

\item \code{marianes_htseq_am3}: Individual replicate from the  a1-3 region

\item \code{marianes_htseq_am4}: Individual replicate from the  CuLFCFe region

\item \code{marianes_htseq_am5}: Individual replicate from the  CuLFCFe region

\item \code{marianes_htseq_am6}: Individual replicate from the  CuLFCFe region

\item \code{marianes_htseq_am7}: Individual replicate from the  p1-4 region

\item \code{marianes_htseq_am8}: Individual replicate from the  p1-4 region

\item \code{marianes_htseq_am9}: Individual replicate from the  p1-4 region

\item \code{marianes_htseq_am10}: Individual replicate from the  a1 region

\item \code{marianes_htseq_am11}: Individual replicate from the  a1 region

\item \code{marianes_htseq_am12}: Individual replicate from the  a1 region

\item \code{marianes_htseq_am13}: Individual replicate from the  a2-3 region

\item \code{marianes_htseq_am14}: Individual replicate from the  a2-3 region

\item \code{marianes_htseq_am15}: Individual replicate from the  a2-3 region

\item \code{marianes_htseq_am16}: Individual replicate from the  Cu region

\item \code{marianes_htseq_am17}: Individual replicate from the  Cu region

\item \code{marianes_htseq_am18}: Individual replicate from the  Cu region

\item \code{marianes_htseq_am19}: Individual replicate from the  LFCFe region

\item \code{marianes_htseq_am20}: Individual replicate from the  LFCFe region

\item \code{marianes_htseq_am21}: Individual replicate from the  LFCFe region

\item \code{marianes_htseq_am22}: Individual replicate from the  Fe region

\item \code{marianes_htseq_am23}: Individual replicate from the  Fe region

\item \code{marianes_htseq_am24}: Individual replicate from the  Fe region

\item \code{marianes_htseq_am25}: Individual replicate from the  p1 region

\item \code{marianes_htseq_am26}: Individual replicate from the  p1 region

\item \code{marianes_htseq_am27}: Individual replicate from the  p1 region

\item \code{marianes_htseq_am28}: Individual replicate from the  p2-4 region

\item \code{marianes_htseq_am29}: Individual replicate from the  p2-4 region

\item \code{marianes_htseq_am30}: Individual replicate from the  p2-4 region
}}

\keyword{datasets}
